lostuff

+++ incs/displayer.hpp
@@ -6,7 +6,7 @@
 /*   By: bguyot <bguyot@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 16:02:58 by bguyot            #+#    #+#             */
-/*   Updated: 2023/11/22 17:33:44 by bguyot           ###   ########.fr       */
+/*   Updated: 2023/11/23 11:28:17 by bguyot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,4 +19,4 @@
 #include <sstream>
 #include <string.h>
 
-void	displayer(float vertices[], unsigned int indices[], unsigned int nb_vertices, unsigned int nb_indices);
+void	displayer(float vertices[], unsigned int indices[], size_t nb_vertices, size_t nb_indices);
